requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==2.0.2
django-contrib-comments==1.8.0
django-multiselectfield==0.1.8
djangorestframework==3.7.7
idna==2.6
ndg-httpsclient==0.4.4
Pillow==5.0.0
pusher==1.7.4
pyasn1==0.4.2
pycparser==2.18
pyOpenSSL==17.5.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
